# openapi.yaml
openapi: 3.0.3
info:
  title: Airport Transfer Marketplace API
  version: 1.1.0
  description: |
    Independent Airport Transfer Marketplace API.
    Covers authentication, suppliers, airports/zones/routes, pricing, bookings, rides, and payments.

    ## Role-Based Access Control

    - `GET /public/*`, `POST /public/*`, `/webhooks/*`: **no authentication**
    - `/auth/*`: no authentication
    - `/supplier/*`: requires JWT with roles: `SUPPLIER_OWNER`, `DISPATCHER`, or `DRIVER`
    - `/admin/*`: requires JWT with role: `ADMIN`

    ## Currency Handling

    - Marketplace operates in **EUR** as base currency
    - Suppliers can set prices in their local currency (TRY, USD, GBP, etc.)
    - Currency conversion happens at booking time using real-time rates
    - Supplier payouts are made in their configured payout currency

    ## Cancellation Policy

    - Free cancellation up to **24 hours** before pickup
    - **50% refund** for cancellations **12–24 hours** before pickup
    - **No refund** for cancellations less than **12 hours** before pickup
    - Supplier cancellations result in full refund + possible compensation
servers:
  - url: https://api.airporttransferportal.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DriverLocationStatus:
      type: string
      enum: [ONLINE, OFFLINE, ON_TRIP, BREAK]
...
    DispatchIssueSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    DispatchIssueStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, ESCALATED, CLOSED]
    DispatchIssueType:
      type: string
      enum: [NO_SHOW_CUSTOMER, NO_SHOW_DRIVER, DRIVER_LATE, VEHICLE_BREAKDOWN, ACCIDENT, CUSTOMER_COMPLAINT, DRIVER_COMPLAINT, PAYMENT_ISSUE, FLIGHT_ISSUE, ADDRESS_ISSUE, OTHER]
...
    MessageSenderType:
      type: string
      enum: [CUSTOMER, DRIVER, DISPATCHER, SYSTEM]
    MessageType:
      type: string
      enum: [TEXT, IMAGE, LOCATION, SYSTEM]
    DriverLocation:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        name:
          type: string
        phone:
          type: string
        supplier:
          type: string
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            heading:
              type: number
            speed:
              type: number
            accuracy:
              type: number
            batteryLevel:
              type: integer
        status:
          $ref: '#/components/schemas/DriverLocationStatus'
        currentRideId:
          type: integer
          nullable: true
        bookingCode:
          type: string
          nullable: true
        lastUpdated:
          type: string
          format: date-time
...
    DispatchIssue:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
          nullable: true
        bookingCode:
          type: string
          nullable: true
        rideId:
          type: integer
          nullable: true
        driverId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        type:
          $ref: '#/components/schemas/DispatchIssueType'
        severity:
          $ref: '#/components/schemas/DispatchIssueSeverity'
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/DispatchIssueStatus'
        resolution:
          type: string
          nullable: true
        reportedBy:
          type: string
        reporterId:
          type: integer
          nullable: true
        resolvedBy:
          type: integer
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customerName:
          type: string
          nullable: true
        driverName:
          type: string
          nullable: true
        supplierName:
          type: string
          nullable: true
...
    Message:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        rideId:
          type: integer
          nullable: true
        senderType:
          $ref: '#/components/schemas/MessageSenderType'
        senderId:
          type: integer
          nullable: true
        senderName:
          type: string
          nullable: true
        message:
          type: string
        messageType:
          $ref: '#/components/schemas/MessageType'
        attachmentUrl:
          type: string
          nullable: true
        location:
          type: object
          nullable: true
          properties:
            lat:
              type: number
            lng:
              type: number
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /dispatch/locations:
    get:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Get driver locations
      description: Returns real-time driver locations with status and current ride info
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DriverLocationStatus'
          description: Filter by driver status
      responses:
        '200':
          description: Driver locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  drivers:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriverLocation'
    post:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Update driver location
      description: Update a driver's GPS location and status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driverId, latitude, longitude]
              properties:
                driverId:
                  type: integer
                latitude:
                  type: number
                longitude:
                  type: number
                heading:
                  type: number
                speed:
                  type: number
                accuracy:
                  type: number
                batteryLevel:
                  type: integer
                status:
                  $ref: '#/components/schemas/DriverLocationStatus'
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /dispatch/issues:
    get:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Get dispatch issues
      description: Returns issues and incidents requiring attention
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, open, in_progress, critical, resolved]
          description: Filter issues by status
      responses:
        '200':
          description: Dispatch issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  issues:
                    type: array
                    items:
                      $ref: '#/components/schemas/DispatchIssue'
    post:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Create dispatch issue
      description: Report a new issue or incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, severity, title]
              properties:
                bookingId:
                  type: integer
                rideId:
                  type: integer
                type:
                  $ref: '#/components/schemas/DispatchIssueType'
                severity:
                  $ref: '#/components/schemas/DispatchIssueSeverity'
                title:
                  type: string
                description:
                  type: string
                reportedBy:
                  type: string
                  enum: [CUSTOMER, DRIVER, DISPATCHER, SYSTEM]
      responses:
        '200':
          description: Issue created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  issueId:
                    type: integer
    put:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Update dispatch issue
      description: Update issue status or resolution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: integer
                status:
                  $ref: '#/components/schemas/DispatchIssueStatus'
                resolution:
                  type: string
                resolvedBy:
                  type: integer
      responses:
        '200':
          description: Issue updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /dispatch/messages/{bookingId}:
    get:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Get messages for booking
      description: Returns message thread for a specific booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
    post:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Send message
      description: Send a message in a booking thread
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, senderType]
              properties:
                message:
                  type: string
                senderType:
                  $ref: '#/components/schemas/MessageSenderType'
                senderName:
                  type: string
                messageType:
                  $ref: '#/components/schemas/MessageType'
                attachmentUrl:
                  type: string
                locationLat:
                  type: number
                locationLng:
                  type: number
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messageId:
                    type: integer
...